extract evaluator
