+++ models/base_model.py
@@ -1,16 +1,12 @@
 import os
 import torch
 import numpy as np
-from collections import OrderedDict
 from abc import ABC, abstractmethod
 from . import networks
 
 from collections import OrderedDict
-from torch.distributions import Categorical
-from util.util import prepare_z_y, one_hot, visualize_imgs 
-from TTUR import fid
-from util.inception import get_inception_score
-from inception_pytorch import inception_utils
+from util.util import visualize_imgs
+
 
 class BaseModel(ABC):
     """This class is an abstract base class (ABC) for models.
@@ -29,7 +25,7 @@ class BaseModel(ABC):
             opt (Option class)-- stores all the experiment flags; needs to be a subclass of BaseOptions
 
         When creating your custom class, you need to implement your own initialization.
-        In this fucntion, you should first call <BaseModel.__init__(self, opt)>
+        In this function, you should first call <BaseModel.__init__(self, opt)>
         Then, you need to define four lists:
             -- self.loss_names (str list):          specify the training losses that you want to plot and save.
             -- self.model_names (str list):         specify the images that you want to display and save.
@@ -39,7 +35,8 @@ class BaseModel(ABC):
         self.opt = opt
         self.gpu_ids = opt.gpu_ids
         self.isTrain = opt.isTrain
-        self.device = torch.device('cuda:{}'.format(self.gpu_ids[0])) if self.gpu_ids else torch.device('cpu')  # get device name: CPU or GPU
+        self.device = torch.device('cuda:{}'.format(self.gpu_ids[0])) if self.gpu_ids else torch.device(
+            'cpu')  # get device name: CPU or GPU
         self.save_dir = os.path.join(opt.checkpoints_dir, opt.name)  # save all the checkpoints to save_dir
         if opt.preprocess != 'scale_width':  # with [scale_width], input images might have different sizes, which hurts the performance of cudnn.benchmark.
             torch.backends.cudnn.benchmark = True
@@ -49,38 +46,6 @@ class BaseModel(ABC):
         self.optimizers = []
         self.image_paths = []
         self.metric = 0  # used for learning rate policy 'plateau'
-
-        # scores init
-        if self.opt.use_pytorch_scores and self.opt.score_name is not None:
-            no_FID = True
-            no_IS = True
-            parallel = len(opt.gpu_ids) > 1 
-            for name in self.opt.score_name:
-                if name == 'FID':
-                    no_FID = False 
-                if name == 'IS':
-                    no_IS = False 
-            self.get_inception_metrics = inception_utils.prepare_inception_metrics(opt.dataset_name, parallel, no_IS, no_FID) 
-        else:
-            for name in self.opt.score_name:
-                if name == 'FID':
-                    STAT_FILE = self.opt.fid_stat_file
-                    INCEPTION_PATH = "./inception_v3/"
-
-                    print("load train stats.. ")
-                    # load precalculated training set statistics
-                    f = np.load(STAT_FILE)
-                    self.mu_real, self.sigma_real = f['mu'][:], f['sigma'][:]
-                    f.close()
-                    print("ok")
-
-                    inception_path = fid.check_or_download_inception(INCEPTION_PATH) # download inception network
-                    fid.create_inception_graph(inception_path)  # load the graph into the current TF graph
-
-                    # config = tf.ConfigProto()
-                    # config.gpu_options.allow_growth = True
-                    # self.sess = tf.Session(config = config)
-                    # self.sess.run(tf.global_variables_initializer())
 
     @staticmethod
     def modify_commandline_options(parser, is_train):
@@ -163,20 +128,13 @@ class BaseModel(ABC):
         lr = self.optimizers[0].param_groups[0]['lr']
         print('learning rate = %.7f' % lr)
 
-    def get_current_visuals(self):
-        """Return visualization images. train.py will display these images with visdom, and save the images to a HTML"""
-        visual_ret = OrderedDict()
-        for name in self.visual_names:
-            if isinstance(name, str):
-                visual_ret[name] = getattr(self, name)
-        return visual_ret
-
     def get_current_losses(self):
-        """Return traning losses / errors. train.py will print out these errors on console, and save them to a file"""
+        """Return training losses / errors. train.py will print out these errors on console, and save them to a file"""
         errors_ret = OrderedDict()
         for name in self.loss_names:
             if isinstance(name, str):
-                errors_ret[name] = float(getattr(self, 'loss_' + name))  # float(...) works for both scalar tensor and float number
+                errors_ret[name] = float(
+                    getattr(self, 'loss_' + name))  # float(...) works for both scalar tensor and float number
         return errors_ret
 
     def save_networks(self, epoch):
@@ -206,7 +164,7 @@ class BaseModel(ABC):
                 if getattr(module, key) is None:
                     state_dict.pop('.'.join(keys))
             if module.__class__.__name__.startswith('InstanceNorm') and \
-               (key == 'num_batches_tracked'):
+                    (key == 'num_batches_tracked'):
                 state_dict.pop('.'.join(keys))
         else:
             self.__patch_instance_norm_state_dict(state_dict, getattr(module, key), keys, i + 1)
@@ -271,8 +229,8 @@ class BaseModel(ABC):
     def get_current_visuals(self):
         if self.opt.model == 'egan':
             # load current best G
-            F = self.Fitness[:,2]
-            idx = np.where(F==max(F))[0][0]
+            F = self.Fitness[:, 2]
+            idx = np.where(F == max(F))[0][0]
             self.netG.load_state_dict(self.G_candis[idx])
 
         visual_ret = OrderedDict()
@@ -290,82 +248,3 @@ class BaseModel(ABC):
             if isinstance(name, str):
                 visual_ret[name] = getattr(self, name)
         return visual_ret
-
-    def get_current_scores(self):
-        if self.opt.model == 'egan':
-            # load current best G
-            F = self.Fitness[:,2]
-            idx = np.where(F==max(F))[0][0]
-            self.netG.load_state_dict(self.G_candis[idx])
-
-        # load current best G
-        scores_ret = OrderedDict()
-
-        samples = torch.zeros((self.opt.evaluation_size, 3, self.opt.crop_size, self.opt.crop_size), device=self.device)
-        n_fid_batches = self.opt.evaluation_size // self.opt.fid_batch_size
-
-        for i in range(n_fid_batches):
-            frm = i * self.opt.fid_batch_size
-            to = frm + self.opt.fid_batch_size
-
-            if self.opt.z_type == 'Gaussian': 
-                z = torch.randn(self.opt.fid_batch_size, self.opt.z_dim, 1, 1, device=self.device)
-            elif self.opt.z_type == 'Uniform': 
-                z = torch.rand(self.opt.fid_batch_size, self.opt.z_dim, 1, 1, device=self.device) *2. - 1.
-
-            if self.opt.cgan:
-                y = self.CatDis.sample([self.opt.fid_batch_size])
-                y = one_hot(y, [self.opt.fid_batch_size])
-
-            if not self.opt.cgan:
-                gen_s = self.netG(z).detach()
-            else:
-                gen_s = self.netG(z, y).detach()
-            samples[frm:to] = gen_s
-            print("\rgenerate fid sample batch %d/%d " % (i + 1, n_fid_batches))
-
-        print("%d samples generating done"%self.opt.evaluation_size)
-
-        if self.opt.use_pytorch_scores:
-            self.IS_mean, self.IS_var, self.FID = self.get_inception_metrics(samples, self.opt.evaluation_size, num_splits=10)
-            if 'FID' in self.opt.score_name:
-                print(self.FID)
-                scores_ret['FID'] = float(self.FID) 
-            if 'IS' in self.opt.score_name:
-                print(self.IS_mean, self.IS_var)
-                scores_ret['IS_mean'] = float(self.IS_mean)
-                scores_ret['IS_var'] = float(self.IS_var)
-
-        else:
-            # Cast, reshape and transpose (BCHW -> BHWC)
-            samples = samples.cpu().numpy()
-            samples = ((samples + 1.0) * 127.5).astype('uint8')
-            samples = samples.reshape(self.opt.evaluation_size, 3, self.opt.crop_size, self.opt.crop_size)
-            samples = samples.transpose(0,2,3,1)
-            for name in self.opt.score_name:
-                if name == 'FID':
-                    mu_gen, sigma_gen = fid.calculate_activation_statistics(samples,
-                                          self.sess,
-                                          batch_size=self.opt.fid_batch_size,
-                                          verbose=True)
-                    print("calculate FID:")
-                    try:
-                        self.FID = fid.calculate_frechet_distance(mu_gen, sigma_gen, self.mu_real, self.sigma_real)
-                    except Exception as e:
-                        print(e)
-                        self.FID=500
-                    print(self.FID)
-                    scores_ret[name] = float(self.FID)
-                if name == 'IS':
-                    Imlist = []
-                    for i in range(len(samples)):
-                        im = samples[i,:,:,:]
-                        Imlist.append(im)
-                    print(np.array(Imlist).shape)
-                    self.IS_mean, self.IS_var = get_inception_score(Imlist)
-
-                    scores_ret['IS_mean'] = float(self.IS_mean)
-                    scores_ret['IS_var'] = float(self.IS_var)
-                    print(self.IS_mean, self.IS_var)
-
-        return scores_ret
